Separate out VM dependencies
