+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: package_config
-version: 0.0.2+4
+version: 0.0.3
 description: Support for working with Package Resolution config files.
 author: Dart Team <misc@dartlang.org>
 homepage: https://github.com/dart-lang/package_config
